% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGEE-package.R
\docType{package}
\name{compGEE}
\alias{compGEE-package}
\alias{compGEE}
\title{compGEE: Longitudinal compositional data analysis via Generalized Estimating Equations}
\description{
The package contains functions for longitudinal compositional data analysis. The method is based on a multiplicative model by Firth and Sammut (2023) to represent data directly on the simplex, without logratio transformations. The longitudinal aspect is modeled via Generalized Estimating Equations (Liang and Zeger, 1986)
}
\references{
Firth, D. and Sammut, F. (2023) Analysis of composition on the original scale of measurement. arXiv preprint arXiv:2312.10548.

Liang, K.-Y. and Zeger, S. L. (1986) Longitudinal data analysis using generalized linear models. Biometrika 73(1), 13–22.
}
